requirements.txt
dj-database-url==0.4.1
Django==1.10.2
django-widget-tweaks==1.4.1
djangorestframework==3.5.3
numpy==1.11.2
psycopg2==2.6.2
requests==2.9.1
virtualenv==15.1.0
waitress==1.0.1
whitenoise==3.2.2
